Last plugins with cfi-s are cleaned from exists checks
Now all plugins with cfi files should have good default values in cfi and no exists() checks.

There are exists() for nested parameters: mappingEntry or PSet.
Example in RecoTau/python/PFTauPrimaryVertexProducer_cfi.py

PFTauPrimaryVertexProducer = cms.EDProducer("PFTauPrimaryVertexProducer",
   ...
   discriminators = cms.VPSet(cms.PSet(discriminator = cms.InputTag('hpsPFTauDiscriminationByDecayModeFinding'),selectionCut = cms.double(0.5))),
)

Which is not a matter of the pluggin system.

Some cases had program logic passed through a parameter:
  if parameter A does not exist - get parameter B etc
  or
  if parameter exists - set something with it, otherwise do nothing.
In these cases one either needs an additional explicit boolean for the logic in the plugin parameters (like loadMVAfromDB)
or a special value for the parameter for "doing nothing" (empty string often corresponds to this in the plugins).
The cases are checked for consistency with current default values.

	modified:   RecoTau/plugins/PATTauDiscriminantCutMultiplexer.cc
	modified:   RecoTau/plugins/PATTauDiscriminationByMVAIsolationRun2.cc
	modified:   RecoTau/plugins/PFRecoTauDiscriminationAgainstMuonMVA.cc
	modified:   RecoTau/plugins/PFRecoTauDiscriminationByMVAIsolation2.cc
	modified:   RecoTau/plugins/PFRecoTauDiscriminationByMVAIsolationRun2.cc
	modified:   RecoTau/plugins/RecoTauCleaner.cc
	modified:   RecoTau/plugins/RecoTauDiscriminantCutMultiplexer.cc
	modified:   RecoTau/python/PATTauDiscriminantCutMultiplexer_cfi.py
	modified:   RecoTau/python/PATTauDiscriminationByMVAIsolationRun2_cff.py
	modified:   RecoTau/python/PFRecoTauDiscriminationAgainstMuonMVA_cfi.py
	modified:   RecoTau/python/PFRecoTauDiscriminationByMVAIsolation2_cff.py
	modified:   RecoTau/python/PFRecoTauDiscriminationByMVAIsolationRun2_cff.py
	modified:   RecoTau/python/RecoTauCleaner_cfi.py
	modified:   RecoTau/python/RecoTauDiscriminantCutMultiplexer_cfi.py