Couple final clean ups
The only remaining ifExists in plugins (not in mappingEntries) are:

1) RecoTau/plugins/RecoTauDiscriminationByFlight.cc
which defines plugin name "PFRecoTauDiscriminationByFlight"
and I don't see it being used anywhere in RecoTauTag/Configuration or RecoTauTag/RecoTau

2) RecoTau/plugins/PFTauPrimaryVertexProducer.cc  (plugin)
3) RecoTau/plugins/RecoTauDiscriminantFromDiscriminator.cc  (pluginFactory)
both have if existsAs<type> which implements a bit of logic and I would like to consult how to handle it better.
Substitute with a more general type, like mappingEntry PSet, or use something else?

(Also there was 1 boolean which I added in a similar situation -- need to check it later, handle the same way with mappingEntry or something else.)

	modified:   RecoTau/plugins/RecoTauDiscriminantFromDiscriminator.cc
	modified:   RecoTau/python/PFTauPrimaryVertexProducer_cfi.py
	modified:   RecoTau/python/RecoTauDiscriminantConfiguration.py